suite-etl/requirements.txt
pyiceberg[pandas,pyarrow,polars]>=0.5.0
boto3>=1.28.0
aiohttp>=3.8.0
python-dotenv>=1.0.0 
fastapi[standard]
web3